rename namespace to defermi_gui
