+++ defermi_gui/ctls.py
@@ -1,27 +1,22 @@
+
 import streamlit as st
 import matplotlib.pyplot as plt
-import matplotlib
 
-from defermi.gui.info import names_info
-from defermi.gui.utils import init_state_variable, download_plot, _get_axis_limits_with_widgets, _filter_names
+from defermi_gui.info import names_info
+from defermi_gui.utils import download_plot, _filter_names, _get_axis_limits_with_widgets
 
 
-class FormationEnergiesPlotter:
+class CTLsPlotter:
 
     def __init__(self,defects_analysis):
         self.da = defects_analysis
 
-    def get_figure(self,entries=None,colors=None,xlim=None,ylim=None):
+    def get_figure(self,entries=None,xlim=None,ylim=None):
         da = self.da
-        if 'chempots' not in st.session_state:
-            st.warning('Chemical potentials are not defined')
-        fig = da.plot_formation_energies(
+        fig = da.plot_ctl(
             entries=entries,
-            chemical_potentials=st.session_state['chempots'],
             figsize=st.session_state['figsize'],
             fontsize=st.session_state['fontsize'],
-            colors=colors,
-            xlim=xlim,
             ylim=ylim)
         fig.grid()
         fig.xlabel(plt.gca().get_xlabel(), fontsize=st.session_state['label_size'])
@@ -30,73 +25,65 @@ class FormationEnergiesPlotter:
         fig = ax.get_figure()
         fig.patch.set_alpha(st.session_state['alpha'])
         ax.patch.set_alpha(st.session_state['alpha'])
-        st.session_state['formation_energies_figure'] = fig
+        st.session_state['ctls_figure'] = fig
         return fig
-
-
+    
     def get_axis_limits(self):
         da = self.da
         set_xlim, xlim = _get_axis_limits_with_widgets(
                                                     label='xlim',
-                                                    key='eform',
+                                                    key='ctl',
                                                     default=(-0.5,da.band_gap+0.5),
                                                     boundaries=(-3.,da.band_gap+3.)) 
         xlim = xlim if set_xlim else None
 
         set_ylim, ylim = _get_axis_limits_with_widgets(
                                                     label='ylim',
-                                                    key='eform',
+                                                    key='ctl',
                                                     default=(-20.,30.),
                                                     boundaries=(-20.,30.))
         ylim = ylim if set_ylim else None
         return xlim,ylim
 
 
-    def get_entries_and_colors(self,filter_names=True):
+    def get_entries(self,filter_names=True):
         da = self.da
         defect_names = da.names
         if filter_names:
-            names = _filter_names(defect_names=defect_names,key='eform')
+            names = _filter_names(defect_names=defect_names,key='ctl')
         else:
             names = defect_names
-
         entries = da.select_entries(names=names)
-        colors = []
-        ordered_names = []
-        for entry in entries:
-            if entry.name not in ordered_names:
-                ordered_names.append(entry.name)      
-        colors = [st.session_state.color_dict[name] for name in ordered_names]
+        return entries
 
-        return entries,colors
+
 
 
 def main():
-    st.set_page_config(layout="wide")
-    st.title('Formation Energies')
 
-    if st.session_state.da and 'chempots' in st.session_state:
+    st.set_page_config(layout="wide")
+    st.title("Charge Transition Levels")
+
+    if st.session_state.da:
         da = st.session_state.da
         cols = st.columns([0.7,0.3])
-        plotter = FormationEnergiesPlotter(defects_analysis=da)
+        plotter = CTLsPlotter(defects_analysis=da)
         with cols[1]:
             xlim,ylim = plotter.get_axis_limits()
-            entries,colors = plotter.get_entries_and_colors(da)
+            entries = plotter.get_entries(da)
 
         with cols[0]:
-            fig = plotter.get_figure(entries,colors,xlim,ylim)
-            st.pyplot(fig, clear_figure=False, width="content")
+            fig = plotter.get_figure(entries,xlim,ylim)
+            st.pyplot(fig, clear_figure=False, width="stretch")
 
         with cols[1]:
             with st.popover(label='ℹ️',help='Info',type='tertiary'):
                 st.write(names_info)
             st.write('')                    
-            download_plot(fig=fig,filename='formation_energies.pdf')
+            download_plot(fig=fig,filename='ctl.pdf')
+
 
 if __name__ == '__main__':
     main()
-
-
-
