+++ defermi_gui/initialize.py
@@ -2,8 +2,8 @@
 import streamlit as st
 
 from defermi import DefectsAnalysis
-from defermi.gui.defaults import get_default_df
-from defermi.gui.utils import init_state_variable
+from defermi_gui.defaults import get_default_df
+from defermi_gui.utils import init_state_variable
 
 
 def initialize_state_variables():
