+++ defermi_gui/presets/vacancies.csv
